Shift parsing of VCF data into VariationFile
