+++ genomics_data_index/storage/io/mutation/variants_processor/VcfVariantsTableProcessor.py
@@ -1,0 +1,15 @@
+import abc
+from typing import Generator, List
+import pandas as pd
+
+from genomics_data_index.storage.io.mutation.NucleotideSampleData import NucleotideSampleData
+
+
+class VcfVariantsTableProcessor(abc.ABC):
+
+    def __init__(self):
+        pass
+
+    @abc.abstractmethod
+    def process(self, sample_data: List[NucleotideSampleData]) -> Generator[pd.DataFrame, None, None]:
+        pass
